+++ cookbooks/mu-activedirectory/metadata.rb
@@ -1,8 +1,8 @@
-name             'active-directory'
+name             'mu-activedirectory'
 maintainer       'eGlobalTech,'
 maintainer_email 'ecap-developers@googlegroups.com'
 license          'All rights reserved'
-description      'Installs/Configures active-directory'
+description      'Installs/Configures mu-activedirectory'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
 depends          "windows"