+++ cookbooks/mu-activedirectory/recipes/default.rb
@@ -1,5 +1,5 @@
 #
-# Cookbook Name:: active-directory
+# Cookbook Name:: mu-activedirectory
 # Recipe:: default
 #
 # Copyright 2015, eGlobalTech,