+++ cookbooks/mu-activedirectory/recipes/domain-node.rb
@@ -1,5 +1,5 @@
 #
-# Cookbook Name:: active-directory
+# Cookbook Name:: mu-activedirectory
 # Recipe:: domain-node
 #
 # Copyright 2015, eGlobalTech,